fix "add-apt-repository: command not found" error
For the "happy path" to makes things work, we use the Ubuntu Cloud
Archives. This repository has to be added. The command to do that
is not available on the gate test nodes [1].

This change installs a package which brings the command
"add-apt-repository" with it.

References:
[1] https://github.com/openstack-infra/project-config/commit/
f90d35885e1143e5a1b67804735fd0c2f294457c

Change-Id: I608c8a35297765283a61ddfadf32d8191186d6c3