Add cleanup of package "software-properties-common"
With commit d126b3f92e56 we added the pgk "software-properties-common"
to allow the installation of the new repositories.
This is a follow up patch which implements the cleanup of this package
in case the devstack-plugin is used outside of a OpenStack testing
job (for example a test VM of developer who doesn't throw away the
VM immediately).

Change-Id: I474341bbba567c8edb42b2b1bf760d09dfb9ddc5